YES Got custom slugs!!
Wow, this is the farthest i have *ever* gotten with a cms of this level of granularity.

Apparently you cant use "await" without a preceding "async." I wrapped the awaits in asyncs and I was good.

This filters out blogposts that are drafts so it only displays the published ones (ie the ones that are not drafts)